requirements.txt
beautifulsoup4==4.14.2
ddgs==9.6.0
langchain==1.0.3
langchain_core==1.0.1
langchain_ollama==1.0.0
langgraph==1.0.1
pandas==2.3.3
psycopg2==2.9.11
python-dotenv==1.2.1
requests==2.32.5
